Refactor project structure: update imports, components, and remove unused files
